ingestion/requirements.txt
kafka-python
vaderSentiment
snowflake.connector 
